Refactor quote columns

+++ database/migrations/2024_03_13_165842_create_quotes_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->foreignIdFor(User::class)->constrained();
             $table->foreignIdFor(Vertical::class)->constrained();
-            $table->decimal('total');
+            $table->bigInteger('total');
             $table->json('items');
             $table->softDeletes();
             $table->timestamps();
