+++ database/migrations/2024_03_14_110911_add_taxes_to_quotes_table.php
@@ -12,8 +12,8 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('quotes', function (Blueprint $table) {
-            $table->decimal('subtotal');
-            $table->decimal('taxes');
+            $table->bigInteger('subtotal');
+            $table->bigInteger('taxes');
         });
     }
